requirements.txt
appdirs==1.4.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
click==6.7
decorator==4.0.11
enum34==1.1.6
Flask==0.12.1
Flask-Caching==1.3.3
futures==3.1.1
gevent==1.2.2
greenlet==0.4.12
itsdangerous==0.24
Jinja2==2.9.6
kazoo==2.2.1
MarkupSafe==1.0
msgpack-python==0.4.8
mysqlclient==1.3.10
packaging==16.8
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pykafka==2.7.0-dev.2
pyparsing==2.2.0
PyYAML==3.12
redis==2.10.6
scandir==1.5
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.9
tabulate==0.8.1
traitlets==4.3.2
trollius==2.1
uWSGI==2.0.15
visitor==0.1.3
wcwidth==0.1.7
Werkzeug==0.12.1
